[Feature] countmvvm (#63)
* [REFACTOR] : ♻️

* [FEAT] : ⚙️ modifie adapter to listadatper

* [FEAT] : ⚙️ add primary key

* [FEAT] : ⚙️ apply the MVVM architecture

* [FEAT] : ⚙️ apply the MVVM architecture

* [DESIGN] : 🎨

* [REFACTOR] : ♻️

* [REFACTOR] : ♻️

* [DESIGN] : 🎨

* [FEAT] : ⚙️ record exercise

* [REFACTOR] : ♻️

* [!HOTFIX] : 💥email_image add

* [REFACTOR] : ♻️

---------

+++ app/src/main/java/com/example/alldayfit/db/model/FirebaseModel.kt
@@ -24,7 +24,7 @@ data class FirebaseModel(
     )
 
     data class ExerciseRecord(
-        val totalTime: Int,
+        val totalTime: Long,
         val logDate: String,
     )
